[MR] From release into master

See merge request s03-final/s03p31a303!243

Former-commit-id: 26d6f501c6f6f23b4b9848e03fbf8c5a6c7e36d5 [formerly 3cdca0b3fc54e9b94212dd8ed104421f80e07f95] [formerly c5918947663c22a58b5a6b151cecd59d181bbba7 [formerly 2d847bee12d60f7a760745f80d1ee3bad36b8ca1]]
Former-commit-id: ccb8f9a4aff4879bf840552d09393cbe18a7d0af [formerly 53ecb23d0f8b2e0456dcbcad542df48f518b7d82]
Former-commit-id: 233b200b65c0d4c6b8958488696f6cc9db973d7b

+++ frontend/RunWithMe/src/views/app/mypages/friends.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="main-content">
-    <breadcumb :page="'Friends List'" :folder="'MyPage'" />
+    <breadcumb :page="'친구 목록'" :folder="'MyPage'" />
     <!-- <div class="wrapper"> -->
     <b-card>
       <vue-good-table
@@ -8,12 +8,13 @@
         :line-numbers="false"
         :search-options="{
           enabled: true,
-          placeholder: 'Search this table'
+          placeholder: '검색'
         }"
         :pagination-options="{
           enabled: true,
           mode: 'records'
         }"
+        rowsPerPageText="d"
         styleClass="tableOne vgt-table"
         :rows="rows"
       >
@@ -44,31 +45,42 @@
 
         <template slot="table-row" slot-scope="props">
           <span v-if="props.column.field == 'button'">
-            <a href="" @click="chatFriend(props.row.userId);">
+            <a @click="chatFriend(props.row.userId.username);">
               <i class="i-Speach-Bubble-8 text-25 text-success mr-2"></i>
               {{ props.row.button }}</a
             >
-            <a href="" @click="deleteFriend(props.row.userId);">
+            <a @click="deleteFriend(props.row.userId.userId);">
               <i class="i-Close-Window text-25 text-danger"></i>
               {{ props.row.button }}</a
             >
           </span>
           <span v-else-if="props.column.field == 'gender'">
-            <div v-if="props.row.gender=='1'">
+            <div v-if="props.row.userId.gender=='1'">
               남자
             </div>
             <div v-else>
               여자
             </div>
           </span>
+          <span class="friendList" v-else-if="props.column.field == 'gugunId.gugunName'">
+            {{props.row.userId.gugunId.sidoId.sidoName}} {{props.row.userId.gugunId.gugunName}}
+          </span>
           <span v-else-if="props.column.field == 'username'">
             <a href="">
-              <div class="ul-widget-app__profile-pic">
-                <img
-                  class="profile-picture avatar-sm mb-2 rounded-circle img-fluid"
-                  :src="props.row.profile"
-                  alt=""/>
-                {{ props.row.username }}
+              <div class="ul-widget-app__profile-pic" style="text-align:center">
+                <div v-if="props.row.userId.profile!=null">
+                  <img
+                    class="profile-picture avatar-sm mb-2 rounded-circle img-fluid"
+                    :src="props.row.userId.profile" alt=""/>
+                </div>
+                <div v-else>
+                  <img
+                    class="profile-picture avatar-sm mb-2 rounded-circle img-fluid"
+                    :src="defaultProfile" alt=""/>
+                </div>
+                  <div class="friendList">
+                    {{ props.row.userId.username }}
+                  </div>
               </div>
             </a>
           </span>
@@ -80,7 +92,18 @@
 
 <script>
 import http from "@/utils/http-common";
-import { mapGetters } from "vuex";
+import { mapGetters,mapMutations } from "vuex";
+
+const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+        confirmButton: 'b-button',
+        cancelButton: 'b-button'
+    },
+    buttonsStyling: {
+        confirmButton: "outline-dark m-1",
+        cancelButton: 'info ripple m-1'
+    }
+})
 
 export default {
   metaInfo: {
@@ -93,19 +116,20 @@ export default {
           label: "이름",
           field: "username"
         },
-        {
-          label: "Email",
-          field: "userEmail"
-        },
+        // {
+        //   label: "Email",
+        //   field: "userEmail"
+        // },
         // {
         //   label: "등급",
         //   field: "span",
         //   html: true
+
         // },
-        {
-          label: "성별",
-          field: "gender"
-        },
+        // {
+        //   label: "성별",
+        //   field: "gender"
+        // },
         {
           label: "지역",
           field: "gugunId.gugunName"
@@ -121,33 +145,96 @@ export default {
     };
   },
   computed: {
-    ...mapGetters(["userInfo"]),
+    ...mapGetters(["userInfo","defaultProfile"]),
   },
   mounted() {
     this.getFriendList()
+    this.$store.commit('closeSidebar')
   },
   methods: {
+    ...mapMutations(["closeSidebar"]),
     chatFriend(friendId){
-      alert(friendId+"친구와 채팅합니다")
+      swalWithBootstrapButtons.fire({
+          title: friendId+'님과 채팅하시겠습니까?',
+          text: "채팅방으로 이동합니다.",
+          icon: 'info',
+          showCancelButton: true,
+          confirmButtonText: '네',
+          cancelButtonText: '아니요',
+          reverseButtons: false
+
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.$router.push("/app/apps/chat");
+        } else if (
+              result.dismiss === Swal.DismissReason.cancel
+          ) {
+              swalWithBootstrapButtons.fire(
+              '취소되었습니다.',
+              '',
+              'error'
+              )
+          }
+      })
     },
     deleteFriend(friendId){
-      http.delete(`friends/friend/${friendId}`, {
-      })
-      .then(data => {
-        alert('친구가 삭제되었습니다!')
+      swalWithBootstrapButtons.fire({
+          title: '팔로우를 취소하시겠습니까?',
+          text: "다시 팔로우 가능합니다!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: '팔로우를 취소할래요.',
+          cancelButtonText: '계속 팔로우할래요.',
+          reverseButtons: true
+
+      }).then((result) => {
+        if (result.isConfirmed) {
+            http.delete(`friends/friend/${friendId}`, {
+            })
+            .then(({data}) => {
+                if(data.status == 200) {
+                    this.isFriend = false;
+                }
+            })
+            .catch((error) => {
+                //console.log(error);
+                return;
+            });
+            swalWithBootstrapButtons.fire(
+              '팔로우 취소',
+              '팔로우가 취소되었습니다.',
+              'success'
+            ).then((result) => {
+              this.$router.go(0);
+            })
+        } else if (
+              /* Read more about handling dismissals below */
+              result.dismiss === Swal.DismissReason.cancel
+          ) {
+              swalWithBootstrapButtons.fire(
+              '취소되었습니다.',
+              '계속 팔로우 중입니다.',
+              'error'
+              )
+          }
       })
     },
     getFriendList(){
-      console.log(this.userInfo.userId)
+      //console.log(this.userInfo.userId)
       // http.get(`friends/contacts/${this.userInfo.userId}`)
       http.get(`friends/contacts`)
       .then((res) => {
-        console.log(res.data.data);
+        //console.log(res.data.data);
         this.rows = res.data.data
     });
     }
   }
 };
 </script>
-<style >
+<style scoped>
+.friendList {
+    overflow:hidden;
+  text-overflow:ellipsis;
+  white-space:nowrap;
+}
 </style>
